Struct aws_sdk_servicecatalog::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for AWS242ServiceCatalogService.

This client allows ergonomic access to a AWS242ServiceCatalogService-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn accept_portfolio_share(&self) -> AcceptPortfolioShare<C, M, R>

pub fn associate_budget_with_resource(
&self
) -> AssociateBudgetWithResource<C, M, R>

pub fn associate_principal_with_portfolio(
&self
) -> AssociatePrincipalWithPortfolio<C, M, R>

pub fn associate_product_with_portfolio(
&self
) -> AssociateProductWithPortfolio<C, M, R>

pub fn associate_service_action_with_provisioning_artifact(
&self
) -> AssociateServiceActionWithProvisioningArtifact<C, M, R>

pub fn associate_tag_option_with_resource(
&self
) -> AssociateTagOptionWithResource<C, M, R>

pub fn batch_associate_service_action_with_provisioning_artifact(
&self
) -> BatchAssociateServiceActionWithProvisioningArtifact<C, M, R>

pub fn batch_disassociate_service_action_from_provisioning_artifact(
&self
) -> BatchDisassociateServiceActionFromProvisioningArtifact<C, M, R>

pub fn copy_product(&self) -> CopyProduct<C, M, R>

pub fn create_constraint(&self) -> CreateConstraint<C, M, R>

pub fn create_portfolio(&self) -> CreatePortfolio<C, M, R>

pub fn create_portfolio_share(&self) -> CreatePortfolioShare<C, M, R>

pub fn create_product(&self) -> CreateProduct<C, M, R>

pub fn create_provisioned_product_plan(
&self
) -> CreateProvisionedProductPlan<C, M, R>

pub fn create_provisioning_artifact(
&self
) -> CreateProvisioningArtifact<C, M, R>

pub fn create_service_action(&self) -> CreateServiceAction<C, M, R>

pub fn create_tag_option(&self) -> CreateTagOption<C, M, R>

pub fn delete_constraint(&self) -> DeleteConstraint<C, M, R>

pub fn delete_portfolio(&self) -> DeletePortfolio<C, M, R>

pub fn delete_portfolio_share(&self) -> DeletePortfolioShare<C, M, R>

pub fn delete_product(&self) -> DeleteProduct<C, M, R>

pub fn delete_provisioned_product_plan(
&self
) -> DeleteProvisionedProductPlan<C, M, R>

pub fn delete_provisioning_artifact(
&self
) -> DeleteProvisioningArtifact<C, M, R>

pub fn delete_service_action(&self) -> DeleteServiceAction<C, M, R>

pub fn delete_tag_option(&self) -> DeleteTagOption<C, M, R>

pub fn describe_constraint(&self) -> DescribeConstraint<C, M, R>

pub fn describe_copy_product_status(&self) -> DescribeCopyProductStatus<C, M, R>

pub fn describe_portfolio(&self) -> DescribePortfolio<C, M, R>

pub fn describe_portfolio_shares(&self) -> DescribePortfolioShares<C, M, R>

pub fn describe_portfolio_share_status(
&self
) -> DescribePortfolioShareStatus<C, M, R>

pub fn describe_product(&self) -> DescribeProduct<C, M, R>

pub fn describe_product_as_admin(&self) -> DescribeProductAsAdmin<C, M, R>

pub fn describe_product_view(&self) -> DescribeProductView<C, M, R>

pub fn describe_provisioned_product(
&self
) -> DescribeProvisionedProduct<C, M, R>

pub fn describe_provisioned_product_plan(
&self
) -> DescribeProvisionedProductPlan<C, M, R>

pub fn describe_provisioning_artifact(
&self
) -> DescribeProvisioningArtifact<C, M, R>

pub fn describe_provisioning_parameters(
&self
) -> DescribeProvisioningParameters<C, M, R>

pub fn describe_record(&self) -> DescribeRecord<C, M, R>

pub fn describe_service_action(&self) -> DescribeServiceAction<C, M, R>

pub fn describe_service_action_execution_parameters(
&self
) -> DescribeServiceActionExecutionParameters<C, M, R>

pub fn describe_tag_option(&self) -> DescribeTagOption<C, M, R>

pub fn disable_aws_organizations_access(
&self
) -> DisableAWSOrganizationsAccess<C, M, R>

pub fn disassociate_budget_from_resource(
&self
) -> DisassociateBudgetFromResource<C, M, R>

pub fn disassociate_principal_from_portfolio(
&self
) -> DisassociatePrincipalFromPortfolio<C, M, R>

pub fn disassociate_product_from_portfolio(
&self
) -> DisassociateProductFromPortfolio<C, M, R>

pub fn disassociate_service_action_from_provisioning_artifact(
&self
) -> DisassociateServiceActionFromProvisioningArtifact<C, M, R>

pub fn disassociate_tag_option_from_resource(
&self
) -> DisassociateTagOptionFromResource<C, M, R>

pub fn enable_aws_organizations_access(
&self
) -> EnableAWSOrganizationsAccess<C, M, R>

pub fn execute_provisioned_product_plan(
&self
) -> ExecuteProvisionedProductPlan<C, M, R>

pub fn execute_provisioned_product_service_action(
&self
) -> ExecuteProvisionedProductServiceAction<C, M, R>

pub fn get_aws_organizations_access_status(
&self
) -> GetAWSOrganizationsAccessStatus<C, M, R>

pub fn get_provisioned_product_outputs(
&self
) -> GetProvisionedProductOutputs<C, M, R>

pub fn import_as_provisioned_product(
&self
) -> ImportAsProvisionedProduct<C, M, R>

pub fn list_accepted_portfolio_shares(
&self
) -> ListAcceptedPortfolioShares<C, M, R>

pub fn list_budgets_for_resource(&self) -> ListBudgetsForResource<C, M, R>

pub fn list_constraints_for_portfolio(
&self
) -> ListConstraintsForPortfolio<C, M, R>

pub fn list_launch_paths(&self) -> ListLaunchPaths<C, M, R>

pub fn list_organization_portfolio_access(
&self
) -> ListOrganizationPortfolioAccess<C, M, R>

pub fn list_portfolio_access(&self) -> ListPortfolioAccess<C, M, R>

pub fn list_portfolios(&self) -> ListPortfolios<C, M, R>

pub fn list_portfolios_for_product(&self) -> ListPortfoliosForProduct<C, M, R>

pub fn list_principals_for_portfolio(
&self
) -> ListPrincipalsForPortfolio<C, M, R>

pub fn list_provisioned_product_plans(
&self
) -> ListProvisionedProductPlans<C, M, R>

pub fn list_provisioning_artifacts(&self) -> ListProvisioningArtifacts<C, M, R>

pub fn list_provisioning_artifacts_for_service_action(
&self
) -> ListProvisioningArtifactsForServiceAction<C, M, R>

pub fn list_record_history(&self) -> ListRecordHistory<C, M, R>

pub fn list_resources_for_tag_option(
&self
) -> ListResourcesForTagOption<C, M, R>

pub fn list_service_actions(&self) -> ListServiceActions<C, M, R>

pub fn list_service_actions_for_provisioning_artifact(
&self
) -> ListServiceActionsForProvisioningArtifact<C, M, R>

pub fn list_stack_instances_for_provisioned_product(
&self
) -> ListStackInstancesForProvisionedProduct<C, M, R>

pub fn list_tag_options(&self) -> ListTagOptions<C, M, R>

pub fn provision_product(&self) -> ProvisionProduct<C, M, R>

pub fn reject_portfolio_share(&self) -> RejectPortfolioShare<C, M, R>

pub fn scan_provisioned_products(&self) -> ScanProvisionedProducts<C, M, R>

pub fn search_products(&self) -> SearchProducts<C, M, R>

pub fn search_products_as_admin(&self) -> SearchProductsAsAdmin<C, M, R>

pub fn search_provisioned_products(&self) -> SearchProvisionedProducts<C, M, R>

pub fn terminate_provisioned_product(
&self
) -> TerminateProvisionedProduct<C, M, R>

pub fn update_constraint(&self) -> UpdateConstraint<C, M, R>

pub fn update_portfolio(&self) -> UpdatePortfolio<C, M, R>

pub fn update_portfolio_share(&self) -> UpdatePortfolioShare<C, M, R>

pub fn update_product(&self) -> UpdateProduct<C, M, R>

pub fn update_provisioned_product(&self) -> UpdateProvisionedProduct<C, M, R>

pub fn update_provisioned_product_properties(
&self
) -> UpdateProvisionedProductProperties<C, M, R>

pub fn update_provisioning_artifact(
&self
) -> UpdateProvisioningArtifact<C, M, R>

pub fn update_service_action(&self) -> UpdateServiceAction<C, M, R>

pub fn update_tag_option(&self) -> UpdateTagOption<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self