Module aws_sdk_servicecatalog::error

Modules

duplicate_resource_exception

See DuplicateResourceException

invalid_parameters_exception

See InvalidParametersException

invalid_state_exception

See InvalidStateException

limit_exceeded_exception

See LimitExceededException

operation_not_supported_exception

See OperationNotSupportedException

resource_in_use_exception

See ResourceInUseException

resource_not_found_exception

See ResourceNotFoundException

tag_option_not_migrated_exception

See TagOptionNotMigratedException

Structs

AcceptPortfolioShareError

AssociateBudgetWithResourceError

AssociatePrincipalWithPortfolioError

AssociateProductWithPortfolioError

AssociateServiceActionWithProvisioningArtifactError

AssociateTagOptionWithResourceError

BatchAssociateServiceActionWithProvisioningArtifactError

BatchDisassociateServiceActionFromProvisioningArtifactError

CopyProductError

CreateConstraintError

CreatePortfolioError

CreatePortfolioShareError

CreateProductError

CreateProvisionedProductPlanError

CreateProvisioningArtifactError

CreateServiceActionError

CreateTagOptionError

DeleteConstraintError

DeletePortfolioError

DeletePortfolioShareError

DeleteProductError

DeleteProvisionedProductPlanError

DeleteProvisioningArtifactError

DeleteServiceActionError

DeleteTagOptionError

DescribeConstraintError

DescribeCopyProductStatusError

DescribePortfolioError

DescribePortfolioShareStatusError

DescribePortfolioSharesError

DescribeProductAsAdminError

DescribeProductError

DescribeProductViewError

DescribeProvisionedProductError

DescribeProvisionedProductPlanError

DescribeProvisioningArtifactError

DescribeProvisioningParametersError

DescribeRecordError

DescribeServiceActionError

DescribeServiceActionExecutionParametersError

DescribeTagOptionError

DisableAWSOrganizationsAccessError

DisassociateBudgetFromResourceError

DisassociatePrincipalFromPortfolioError

DisassociateProductFromPortfolioError

DisassociateServiceActionFromProvisioningArtifactError

DisassociateTagOptionFromResourceError

DuplicateResourceException

The specified resource is a duplicate.

EnableAWSOrganizationsAccessError

ExecuteProvisionedProductPlanError

ExecuteProvisionedProductServiceActionError

GetAWSOrganizationsAccessStatusError

GetProvisionedProductOutputsError

ImportAsProvisionedProductError

InvalidParametersException

One or more parameters provided to the operation are not valid.

InvalidStateException

An attempt was made to modify a resource that is in a state that is not valid.

LimitExceededException

The current limits of the service would have been exceeded by this operation. Decrease your

ListAcceptedPortfolioSharesError

ListBudgetsForResourceError

ListConstraintsForPortfolioError

ListLaunchPathsError

ListOrganizationPortfolioAccessError

ListPortfolioAccessError

ListPortfoliosError

ListPortfoliosForProductError

ListPrincipalsForPortfolioError

ListProvisionedProductPlansError

ListProvisioningArtifactsError

ListProvisioningArtifactsForServiceActionError

ListRecordHistoryError

ListResourcesForTagOptionError

ListServiceActionsError

ListServiceActionsForProvisioningArtifactError

ListStackInstancesForProvisionedProductError

ListTagOptionsError

OperationNotSupportedException

The operation is not supported.

ProvisionProductError

RejectPortfolioShareError

ResourceInUseException

A resource that is currently in use. Ensure that the resource is not in use and retry the operation.

ResourceNotFoundException

The specified resource was not found.

ScanProvisionedProductsError

SearchProductsAsAdminError

SearchProductsError

SearchProvisionedProductsError

TagOptionNotMigratedException

An operation requiring TagOptions failed because the TagOptions migration process has

TerminateProvisionedProductError

UpdateConstraintError

UpdatePortfolioError

UpdatePortfolioShareError

UpdateProductError

UpdateProvisionedProductError

UpdateProvisionedProductPropertiesError

UpdateProvisioningArtifactError

UpdateServiceActionError

UpdateTagOptionError

Enums

AcceptPortfolioShareErrorKind

AssociateBudgetWithResourceErrorKind

AssociatePrincipalWithPortfolioErrorKind

AssociateProductWithPortfolioErrorKind

AssociateServiceActionWithProvisioningArtifactErrorKind

AssociateTagOptionWithResourceErrorKind

BatchAssociateServiceActionWithProvisioningArtifactErrorKind

BatchDisassociateServiceActionFromProvisioningArtifactErrorKind

CopyProductErrorKind

CreateConstraintErrorKind

CreatePortfolioErrorKind

CreatePortfolioShareErrorKind

CreateProductErrorKind

CreateProvisionedProductPlanErrorKind

CreateProvisioningArtifactErrorKind

CreateServiceActionErrorKind

CreateTagOptionErrorKind

DeleteConstraintErrorKind

DeletePortfolioErrorKind

DeletePortfolioShareErrorKind

DeleteProductErrorKind

DeleteProvisionedProductPlanErrorKind

DeleteProvisioningArtifactErrorKind

DeleteServiceActionErrorKind

DeleteTagOptionErrorKind

DescribeConstraintErrorKind

DescribeCopyProductStatusErrorKind

DescribePortfolioErrorKind

DescribePortfolioShareStatusErrorKind

DescribePortfolioSharesErrorKind

DescribeProductAsAdminErrorKind

DescribeProductErrorKind

DescribeProductViewErrorKind

DescribeProvisionedProductErrorKind

DescribeProvisionedProductPlanErrorKind

DescribeProvisioningArtifactErrorKind

DescribeProvisioningParametersErrorKind

DescribeRecordErrorKind

DescribeServiceActionErrorKind

DescribeServiceActionExecutionParametersErrorKind

DescribeTagOptionErrorKind

DisableAWSOrganizationsAccessErrorKind

DisassociateBudgetFromResourceErrorKind

DisassociatePrincipalFromPortfolioErrorKind

DisassociateProductFromPortfolioErrorKind

DisassociateServiceActionFromProvisioningArtifactErrorKind

DisassociateTagOptionFromResourceErrorKind

EnableAWSOrganizationsAccessErrorKind

ExecuteProvisionedProductPlanErrorKind

ExecuteProvisionedProductServiceActionErrorKind

GetAWSOrganizationsAccessStatusErrorKind

GetProvisionedProductOutputsErrorKind

ImportAsProvisionedProductErrorKind

ListAcceptedPortfolioSharesErrorKind

ListBudgetsForResourceErrorKind

ListConstraintsForPortfolioErrorKind

ListLaunchPathsErrorKind

ListOrganizationPortfolioAccessErrorKind

ListPortfolioAccessErrorKind

ListPortfoliosErrorKind

ListPortfoliosForProductErrorKind

ListPrincipalsForPortfolioErrorKind

ListProvisionedProductPlansErrorKind

ListProvisioningArtifactsErrorKind

ListProvisioningArtifactsForServiceActionErrorKind

ListRecordHistoryErrorKind

ListResourcesForTagOptionErrorKind

ListServiceActionsErrorKind

ListServiceActionsForProvisioningArtifactErrorKind

ListStackInstancesForProvisionedProductErrorKind

ListTagOptionsErrorKind

ProvisionProductErrorKind

RejectPortfolioShareErrorKind

ScanProvisionedProductsErrorKind

SearchProductsAsAdminErrorKind

SearchProductsErrorKind

SearchProvisionedProductsErrorKind

TerminateProvisionedProductErrorKind

UpdateConstraintErrorKind

UpdatePortfolioErrorKind

UpdatePortfolioShareErrorKind

UpdateProductErrorKind

UpdateProvisionedProductErrorKind

UpdateProvisionedProductPropertiesErrorKind

UpdateProvisioningArtifactErrorKind

UpdateServiceActionErrorKind

UpdateTagOptionErrorKind