Module aws_sdk_servicecatalog::model

Modules

access_level_filter

See AccessLevelFilter

budget_detail

See BudgetDetail

cloud_watch_dashboard

See CloudWatchDashboard

constraint_detail

See ConstraintDetail

constraint_summary

See ConstraintSummary

execution_parameter

See ExecutionParameter

failed_service_action_association

See FailedServiceActionAssociation

launch_path

See LaunchPath

launch_path_summary

See LaunchPathSummary

list_record_history_search_filter

See ListRecordHistorySearchFilter

list_tag_options_filters

See ListTagOptionsFilters

organization_node

See OrganizationNode

parameter_constraints

See ParameterConstraints

portfolio_detail

See PortfolioDetail

portfolio_share_detail

See PortfolioShareDetail

principal

See Principal

product_view_aggregation_value

See ProductViewAggregationValue

product_view_detail

See ProductViewDetail

product_view_summary

See ProductViewSummary

provisioned_product_attribute

See ProvisionedProductAttribute

provisioned_product_detail

See ProvisionedProductDetail

provisioned_product_plan_details

See ProvisionedProductPlanDetails

provisioned_product_plan_summary

See ProvisionedProductPlanSummary

provisioning_artifact

See ProvisioningArtifact

provisioning_artifact_detail

See ProvisioningArtifactDetail

provisioning_artifact_output

See ProvisioningArtifactOutput

provisioning_artifact_parameter

See ProvisioningArtifactParameter

provisioning_artifact_preferences

See ProvisioningArtifactPreferences

provisioning_artifact_properties

See ProvisioningArtifactProperties

provisioning_artifact_summary

See ProvisioningArtifactSummary

provisioning_artifact_view

See ProvisioningArtifactView

provisioning_parameter

See ProvisioningParameter

provisioning_preferences

See ProvisioningPreferences

record_detail

See RecordDetail

record_error

See RecordError

record_output

See RecordOutput

record_tag

See RecordTag

resource_change

See ResourceChange

resource_change_detail

See ResourceChangeDetail

resource_detail

See ResourceDetail

resource_target_definition

See ResourceTargetDefinition

service_action_association

See ServiceActionAssociation

service_action_detail

See ServiceActionDetail

service_action_summary

See ServiceActionSummary

share_details

See ShareDetails

share_error

See ShareError

stack_instance

See StackInstance

tag

See Tag

tag_option_detail

See TagOptionDetail

tag_option_summary

See TagOptionSummary

update_provisioning_parameter

See UpdateProvisioningParameter

update_provisioning_preferences

See UpdateProvisioningPreferences

usage_instruction

See UsageInstruction

Structs

AccessLevelFilter

The access level to use to filter results.

BudgetDetail

Information about a budget.

CloudWatchDashboard

Information about a CloudWatch dashboard.

ConstraintDetail

Information about a constraint.

ConstraintSummary

Summary information about a constraint.

ExecutionParameter

Details of an execution parameter value that is passed to a self-service action when executed on a provisioned product.

FailedServiceActionAssociation

An object containing information about the error, along with identifying information about the self-service action and its associations.

LaunchPath

A launch path object.

LaunchPathSummary

Summary information about a product path for a user.

ListRecordHistorySearchFilter

The search filter to use when listing history records.

ListTagOptionsFilters

Filters to use when listing TagOptions.

OrganizationNode

Information about the organization node.

ParameterConstraints

The constraints that the administrator has put on the parameter.

PortfolioDetail

Information about a portfolio.

PortfolioShareDetail

Information about the portfolio share.

Principal

Information about a principal.

ProductViewAggregationValue

A single product view aggregation value/count pair, containing metadata about each

ProductViewDetail

Information about a product view.

ProductViewSummary

Summary information about a product view.

ProvisionedProductAttribute

Information about a provisioned product.

ProvisionedProductDetail

Information about a provisioned product.

ProvisionedProductPlanDetails

Information about a plan.

ProvisionedProductPlanSummary

Summary information about a plan.

ProvisioningArtifact

Information about a provisioning artifact. A provisioning artifact is also known as a product version.

ProvisioningArtifactDetail

Information about a provisioning artifact (also known as a version) for a product.

ProvisioningArtifactOutput

Provisioning artifact output.

ProvisioningArtifactParameter

Information about a parameter used to provision a product.

ProvisioningArtifactPreferences

The user-defined preferences that will be applied during product provisioning, unless overridden by ProvisioningPreferences or UpdateProvisioningPreferences.

ProvisioningArtifactProperties

Information about a provisioning artifact (also known as a version) for a product.

ProvisioningArtifactSummary

Summary information about a provisioning artifact (also known as a version) for a product.

ProvisioningArtifactView

An object that contains summary information about a product view and a provisioning artifact.

ProvisioningParameter

Information about a parameter used to provision a product.

ProvisioningPreferences

The user-defined preferences that will be applied when updating a provisioned

RecordDetail

Information about a request operation.

RecordError

The error code and description resulting from an operation.

RecordOutput

The output for the product created as the result of a request. For example, the output for

RecordTag

Information about a tag, which is a key-value pair.

ResourceChange

Information about a resource change that will occur when a plan is executed.

ResourceChangeDetail

Information about a change to a resource attribute.

ResourceDetail

Information about a resource.

ResourceTargetDefinition

Information about a change to a resource attribute.

ServiceActionAssociation

A self-service action association consisting of the Action ID, the Product ID, and the Provisioning Artifact ID.

ServiceActionDetail

An object containing detailed information about the self-service action.

ServiceActionSummary

Detailed information about the self-service action.

ShareDetails

Information about the portfolio share operation.

ShareError

Errors that occurred during the portfolio share operation.

StackInstance

An AWS CloudFormation stack, in a specific account and region, that's part of a stack set operation. A stack instance is a reference to an attempted or actual stack in a given account within a given region. A stack instance can exist without a stack—for example, if the stack couldn't be created for some reason. A stack instance is associated with only one stack set. Each stack instance contains the ID of its associated stack set, as well as the ID of the actual stack and the stack status.

Tag

Information about a tag. A tag is a key-value pair. Tags are propagated

TagOptionDetail

Information about a TagOption.

TagOptionSummary

Summary information about a TagOption.

UpdateProvisioningParameter

The parameter key-value pair used to update a provisioned product.

UpdateProvisioningPreferences

The user-defined preferences that will be applied when updating a provisioned product. Not all preferences are applicable to all provisioned product types.

UsageInstruction

Additional information provided by the administrator.

Enums

AccessLevelFilterKey

AccessStatus

ChangeAction

CopyOption

CopyProductStatus

DescribePortfolioShareType

EvaluationType

OrganizationNodeType

PortfolioShareType

PrincipalType

ProductSource

ProductType

ProductViewFilterBy

ProductViewSortBy

PropertyKey

ProvisionedProductPlanStatus

ProvisionedProductPlanType

ProvisionedProductStatus

ProvisionedProductViewFilterBy

ProvisioningArtifactGuidance

ProvisioningArtifactPropertyName

ProvisioningArtifactType

RecordStatus

Replacement

RequiresRecreation

ResourceAttribute

ServiceActionAssociationErrorCode

ServiceActionDefinitionKey

ServiceActionDefinitionType

ShareStatus

SortOrder

StackInstanceStatus

StackSetOperationType

Status