Module aws_sdk_servicecatalog::operation

Structs

AcceptPortfolioShare

Accepts an offer to share the specified portfolio.

AssociateBudgetWithResource

Associates the specified budget with the specified resource.

AssociatePrincipalWithPortfolio

Associates the specified principal ARN with the specified portfolio.

AssociateProductWithPortfolio

Associates the specified product with the specified portfolio.

AssociateServiceActionWithProvisioningArtifact

Associates a self-service action with a provisioning artifact.

AssociateTagOptionWithResource

Associate the specified TagOption with the specified portfolio or product.

BatchAssociateServiceActionWithProvisioningArtifact

Associates multiple self-service actions with provisioning artifacts.

BatchDisassociateServiceActionFromProvisioningArtifact

Disassociates a batch of self-service actions from the specified provisioning artifact.

CopyProduct

Copies the specified source product to the specified target product or a new product.

CreateConstraint

Creates a constraint.

CreatePortfolio

Creates a portfolio.

CreatePortfolioShare

Shares the specified portfolio with the specified account or organization node.

CreateProduct

Creates a product.

CreateProvisionedProductPlan

Creates a plan. A plan includes the list of resources to be

CreateProvisioningArtifact

Creates a provisioning artifact (also known as a version) for the specified product.

CreateServiceAction

Creates a self-service action.

CreateTagOption

Creates a TagOption.

DeleteConstraint

Deletes the specified constraint.

DeletePortfolio

Deletes the specified portfolio.

DeletePortfolioShare

Stops sharing the specified portfolio with the specified account or organization

DeleteProduct

Deletes the specified product.

DeleteProvisionedProductPlan

Deletes the specified plan.

DeleteProvisioningArtifact

Deletes the specified provisioning artifact (also known as a version) for the specified product.

DeleteServiceAction

Deletes a self-service action.

DeleteTagOption

Deletes the specified TagOption.

DescribeConstraint

Gets information about the specified constraint.

DescribeCopyProductStatus

Gets the status of the specified copy product operation.

DescribePortfolio

Gets information about the specified portfolio.

DescribePortfolioShareStatus

Gets the status of the specified portfolio share operation. This API can only be called

DescribePortfolioShares

Returns a summary of each of the portfolio shares that were created for the specified portfolio.

DescribeProduct

Gets information about the specified product.

DescribeProductAsAdmin

Gets information about the specified product. This operation is run with administrator access.

DescribeProductView

Gets information about the specified product.

DescribeProvisionedProduct

Gets information about the specified provisioned product.

DescribeProvisionedProductPlan

Gets information about the resource changes for the specified plan.

DescribeProvisioningArtifact

Gets information about the specified provisioning artifact (also known as a version) for the specified product.

DescribeProvisioningParameters

Gets information about the configuration required to provision the specified product using

DescribeRecord

Gets information about the specified request operation.

DescribeServiceAction

Describes a self-service action.

DescribeServiceActionExecutionParameters

Finds the default parameters for a specific self-service action on a specific provisioned product and returns a map of the results to the user.

DescribeTagOption

Gets information about the specified TagOption.

DisableAWSOrganizationsAccess

Disable portfolio sharing through AWS Organizations feature. This feature will not

DisassociateBudgetFromResource

Disassociates the specified budget from the specified resource.

DisassociatePrincipalFromPortfolio

Disassociates a previously associated principal ARN from a specified

DisassociateProductFromPortfolio

Disassociates the specified product from the specified portfolio.

DisassociateServiceActionFromProvisioningArtifact

Disassociates the specified self-service action association from the specified provisioning artifact.

DisassociateTagOptionFromResource

Disassociates the specified TagOption from the specified resource.

EnableAWSOrganizationsAccess

Enable portfolio sharing feature through AWS Organizations. This API will allow Service

ExecuteProvisionedProductPlan

Provisions or modifies a product based on the resource changes for the specified plan.

ExecuteProvisionedProductServiceAction

Executes a self-service action against a provisioned product.

GetAWSOrganizationsAccessStatus

Get the Access Status for AWS Organization portfolio share feature. This API can only be

GetProvisionedProductOutputs

This API takes either a ProvisonedProductId or a ProvisionedProductName, along with a list of one or more output keys, and responds with the key/value pairs of those outputs.

ImportAsProvisionedProduct

Requests the import of a resource as a Service Catalog provisioned product that is

ListAcceptedPortfolioShares

Lists all portfolios for which sharing was accepted by this account.

ListBudgetsForResource

Lists all the budgets associated to the specified resource.

ListConstraintsForPortfolio

Lists the constraints for the specified portfolio and product.

ListLaunchPaths

Lists the paths to the specified product. A path is how the user

ListOrganizationPortfolioAccess

Lists the organization nodes that have access to the specified portfolio. This API can

ListPortfolioAccess

Lists the account IDs that have access to the specified portfolio.

ListPortfolios

Lists all portfolios in the catalog.

ListPortfoliosForProduct

Lists all portfolios that the specified product is associated with.

ListPrincipalsForPortfolio

Lists all principal ARNs associated with the specified portfolio.

ListProvisionedProductPlans

Lists the plans for the specified provisioned product or all plans to which the user has access.

ListProvisioningArtifacts

Lists all provisioning artifacts (also known as versions) for the specified product.

ListProvisioningArtifactsForServiceAction

Lists all provisioning artifacts (also known as versions) for the specified self-service action.

ListRecordHistory

Lists the specified requests or all performed requests.

ListResourcesForTagOption

Lists the resources associated with the specified TagOption.

ListServiceActions

Lists all self-service actions.

ListServiceActionsForProvisioningArtifact

Returns a paginated list of self-service actions associated with the specified Product ID and Provisioning Artifact ID.

ListStackInstancesForProvisionedProduct

Returns summary information about stack instances that are associated with the specified CFN_STACKSET type provisioned product. You can filter for stack instances that are associated with a specific AWS account name or region.

ListTagOptions

Lists the specified TagOptions or all TagOptions.

ProvisionProduct

Provisions the specified product.

RejectPortfolioShare

Rejects an offer to share the specified portfolio.

ScanProvisionedProducts

Lists the provisioned products that are available (not terminated).

SearchProducts

Gets information about the products to which the caller has access.

SearchProductsAsAdmin

Gets information about the products for the specified portfolio or all products.

SearchProvisionedProducts

Gets information about the provisioned products that meet the specified criteria.

TerminateProvisionedProduct

Terminates the specified provisioned product.

UpdateConstraint

Updates the specified constraint.

UpdatePortfolio

Updates the specified portfolio.

UpdatePortfolioShare

Updates the specified portfolio share. You can use this API to enable or disable TagOptions sharing for an existing portfolio share.

UpdateProduct

Updates the specified product.

UpdateProvisionedProduct

Requests updates to the configuration of the specified provisioned product.

UpdateProvisionedProductProperties

Requests updates to the properties of the specified provisioned product.

UpdateProvisioningArtifact

Updates the specified provisioning artifact (also known as a version) for the specified product.

UpdateServiceAction

Updates a self-service action.

UpdateTagOption

Updates the specified TagOption.