Struct aws_sdk_servicecatalog::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for AWS242ServiceCatalogService.

This client allows ergonomic access to a AWS242ServiceCatalogService-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

Creates a client with the given service configuration.

pub fn conf(&self) -> &Config

Returns the client’s configuration.

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn accept_portfolio_share(&self) -> AcceptPortfolioShare<C, M, R>

Constructs a fluent builder for the AcceptPortfolioShare operation.

See AcceptPortfolioShare for more information about the operation and its arguments.

pub fn associate_budget_with_resource(
&self
) -> AssociateBudgetWithResource<C, M, R>

Constructs a fluent builder for the AssociateBudgetWithResource operation.

See AssociateBudgetWithResource for more information about the operation and its arguments.

pub fn associate_principal_with_portfolio(
&self
) -> AssociatePrincipalWithPortfolio<C, M, R>

Constructs a fluent builder for the AssociatePrincipalWithPortfolio operation.

See AssociatePrincipalWithPortfolio for more information about the operation and its arguments.

pub fn associate_product_with_portfolio(
&self
) -> AssociateProductWithPortfolio<C, M, R>

Constructs a fluent builder for the AssociateProductWithPortfolio operation.

See AssociateProductWithPortfolio for more information about the operation and its arguments.

pub fn associate_service_action_with_provisioning_artifact(
&self
) -> AssociateServiceActionWithProvisioningArtifact<C, M, R>

Constructs a fluent builder for the AssociateServiceActionWithProvisioningArtifact operation.

See AssociateServiceActionWithProvisioningArtifact for more information about the operation and its arguments.

pub fn associate_tag_option_with_resource(
&self
) -> AssociateTagOptionWithResource<C, M, R>

Constructs a fluent builder for the AssociateTagOptionWithResource operation.

See AssociateTagOptionWithResource for more information about the operation and its arguments.

pub fn batch_associate_service_action_with_provisioning_artifact(
&self
) -> BatchAssociateServiceActionWithProvisioningArtifact<C, M, R>

Constructs a fluent builder for the BatchAssociateServiceActionWithProvisioningArtifact operation.

See BatchAssociateServiceActionWithProvisioningArtifact for more information about the operation and its arguments.

pub fn batch_disassociate_service_action_from_provisioning_artifact(
&self
) -> BatchDisassociateServiceActionFromProvisioningArtifact<C, M, R>

Constructs a fluent builder for the BatchDisassociateServiceActionFromProvisioningArtifact operation.

See BatchDisassociateServiceActionFromProvisioningArtifact for more information about the operation and its arguments.

pub fn copy_product(&self) -> CopyProduct<C, M, R>

Constructs a fluent builder for the CopyProduct operation.

See CopyProduct for more information about the operation and its arguments.

pub fn create_constraint(&self) -> CreateConstraint<C, M, R>

Constructs a fluent builder for the CreateConstraint operation.

See CreateConstraint for more information about the operation and its arguments.

pub fn create_portfolio(&self) -> CreatePortfolio<C, M, R>

Constructs a fluent builder for the CreatePortfolio operation.

See CreatePortfolio for more information about the operation and its arguments.

pub fn create_portfolio_share(&self) -> CreatePortfolioShare<C, M, R>

Constructs a fluent builder for the CreatePortfolioShare operation.

See CreatePortfolioShare for more information about the operation and its arguments.

pub fn create_product(&self) -> CreateProduct<C, M, R>

Constructs a fluent builder for the CreateProduct operation.

See CreateProduct for more information about the operation and its arguments.

pub fn create_provisioned_product_plan(
&self
) -> CreateProvisionedProductPlan<C, M, R>

Constructs a fluent builder for the CreateProvisionedProductPlan operation.

See CreateProvisionedProductPlan for more information about the operation and its arguments.

pub fn create_provisioning_artifact(
&self
) -> CreateProvisioningArtifact<C, M, R>

Constructs a fluent builder for the CreateProvisioningArtifact operation.

See CreateProvisioningArtifact for more information about the operation and its arguments.

pub fn create_service_action(&self) -> CreateServiceAction<C, M, R>

Constructs a fluent builder for the CreateServiceAction operation.

See CreateServiceAction for more information about the operation and its arguments.

pub fn create_tag_option(&self) -> CreateTagOption<C, M, R>

Constructs a fluent builder for the CreateTagOption operation.

See CreateTagOption for more information about the operation and its arguments.

pub fn delete_constraint(&self) -> DeleteConstraint<C, M, R>

Constructs a fluent builder for the DeleteConstraint operation.

See DeleteConstraint for more information about the operation and its arguments.

pub fn delete_portfolio(&self) -> DeletePortfolio<C, M, R>

Constructs a fluent builder for the DeletePortfolio operation.

See DeletePortfolio for more information about the operation and its arguments.

pub fn delete_portfolio_share(&self) -> DeletePortfolioShare<C, M, R>

Constructs a fluent builder for the DeletePortfolioShare operation.

See DeletePortfolioShare for more information about the operation and its arguments.

pub fn delete_product(&self) -> DeleteProduct<C, M, R>

Constructs a fluent builder for the DeleteProduct operation.

See DeleteProduct for more information about the operation and its arguments.

pub fn delete_provisioned_product_plan(
&self
) -> DeleteProvisionedProductPlan<C, M, R>

Constructs a fluent builder for the DeleteProvisionedProductPlan operation.

See DeleteProvisionedProductPlan for more information about the operation and its arguments.

pub fn delete_provisioning_artifact(
&self
) -> DeleteProvisioningArtifact<C, M, R>

Constructs a fluent builder for the DeleteProvisioningArtifact operation.

See DeleteProvisioningArtifact for more information about the operation and its arguments.

pub fn delete_service_action(&self) -> DeleteServiceAction<C, M, R>

Constructs a fluent builder for the DeleteServiceAction operation.

See DeleteServiceAction for more information about the operation and its arguments.

pub fn delete_tag_option(&self) -> DeleteTagOption<C, M, R>

Constructs a fluent builder for the DeleteTagOption operation.

See DeleteTagOption for more information about the operation and its arguments.

pub fn describe_constraint(&self) -> DescribeConstraint<C, M, R>

Constructs a fluent builder for the DescribeConstraint operation.

See DescribeConstraint for more information about the operation and its arguments.

pub fn describe_copy_product_status(&self) -> DescribeCopyProductStatus<C, M, R>

Constructs a fluent builder for the DescribeCopyProductStatus operation.

See DescribeCopyProductStatus for more information about the operation and its arguments.

pub fn describe_portfolio(&self) -> DescribePortfolio<C, M, R>

Constructs a fluent builder for the DescribePortfolio operation.

See DescribePortfolio for more information about the operation and its arguments.

pub fn describe_portfolio_shares(&self) -> DescribePortfolioShares<C, M, R>

Constructs a fluent builder for the DescribePortfolioShares operation.

See DescribePortfolioShares for more information about the operation and its arguments.

pub fn describe_portfolio_share_status(
&self
) -> DescribePortfolioShareStatus<C, M, R>

Constructs a fluent builder for the DescribePortfolioShareStatus operation.

See DescribePortfolioShareStatus for more information about the operation and its arguments.

pub fn describe_product(&self) -> DescribeProduct<C, M, R>

Constructs a fluent builder for the DescribeProduct operation.

See DescribeProduct for more information about the operation and its arguments.

pub fn describe_product_as_admin(&self) -> DescribeProductAsAdmin<C, M, R>

Constructs a fluent builder for the DescribeProductAsAdmin operation.

See DescribeProductAsAdmin for more information about the operation and its arguments.

pub fn describe_product_view(&self) -> DescribeProductView<C, M, R>

Constructs a fluent builder for the DescribeProductView operation.

See DescribeProductView for more information about the operation and its arguments.

pub fn describe_provisioned_product(
&self
) -> DescribeProvisionedProduct<C, M, R>

Constructs a fluent builder for the DescribeProvisionedProduct operation.

See DescribeProvisionedProduct for more information about the operation and its arguments.

pub fn describe_provisioned_product_plan(
&self
) -> DescribeProvisionedProductPlan<C, M, R>

Constructs a fluent builder for the DescribeProvisionedProductPlan operation.

See DescribeProvisionedProductPlan for more information about the operation and its arguments.

pub fn describe_provisioning_artifact(
&self
) -> DescribeProvisioningArtifact<C, M, R>

Constructs a fluent builder for the DescribeProvisioningArtifact operation.

See DescribeProvisioningArtifact for more information about the operation and its arguments.

pub fn describe_provisioning_parameters(
&self
) -> DescribeProvisioningParameters<C, M, R>

Constructs a fluent builder for the DescribeProvisioningParameters operation.

See DescribeProvisioningParameters for more information about the operation and its arguments.

pub fn describe_record(&self) -> DescribeRecord<C, M, R>

Constructs a fluent builder for the DescribeRecord operation.

See DescribeRecord for more information about the operation and its arguments.

pub fn describe_service_action(&self) -> DescribeServiceAction<C, M, R>

Constructs a fluent builder for the DescribeServiceAction operation.

See DescribeServiceAction for more information about the operation and its arguments.

pub fn describe_service_action_execution_parameters(
&self
) -> DescribeServiceActionExecutionParameters<C, M, R>

Constructs a fluent builder for the DescribeServiceActionExecutionParameters operation.

See DescribeServiceActionExecutionParameters for more information about the operation and its arguments.

pub fn describe_tag_option(&self) -> DescribeTagOption<C, M, R>

Constructs a fluent builder for the DescribeTagOption operation.

See DescribeTagOption for more information about the operation and its arguments.

pub fn disable_aws_organizations_access(
&self
) -> DisableAWSOrganizationsAccess<C, M, R>

Constructs a fluent builder for the DisableAWSOrganizationsAccess operation.

See DisableAWSOrganizationsAccess for more information about the operation and its arguments.

pub fn disassociate_budget_from_resource(
&self
) -> DisassociateBudgetFromResource<C, M, R>

Constructs a fluent builder for the DisassociateBudgetFromResource operation.

See DisassociateBudgetFromResource for more information about the operation and its arguments.

pub fn disassociate_principal_from_portfolio(
&self
) -> DisassociatePrincipalFromPortfolio<C, M, R>

Constructs a fluent builder for the DisassociatePrincipalFromPortfolio operation.

See DisassociatePrincipalFromPortfolio for more information about the operation and its arguments.

pub fn disassociate_product_from_portfolio(
&self
) -> DisassociateProductFromPortfolio<C, M, R>

Constructs a fluent builder for the DisassociateProductFromPortfolio operation.

See DisassociateProductFromPortfolio for more information about the operation and its arguments.

pub fn disassociate_service_action_from_provisioning_artifact(
&self
) -> DisassociateServiceActionFromProvisioningArtifact<C, M, R>

Constructs a fluent builder for the DisassociateServiceActionFromProvisioningArtifact operation.

See DisassociateServiceActionFromProvisioningArtifact for more information about the operation and its arguments.

pub fn disassociate_tag_option_from_resource(
&self
) -> DisassociateTagOptionFromResource<C, M, R>

Constructs a fluent builder for the DisassociateTagOptionFromResource operation.

See DisassociateTagOptionFromResource for more information about the operation and its arguments.

pub fn enable_aws_organizations_access(
&self
) -> EnableAWSOrganizationsAccess<C, M, R>

Constructs a fluent builder for the EnableAWSOrganizationsAccess operation.

See EnableAWSOrganizationsAccess for more information about the operation and its arguments.

pub fn execute_provisioned_product_plan(
&self
) -> ExecuteProvisionedProductPlan<C, M, R>

Constructs a fluent builder for the ExecuteProvisionedProductPlan operation.

See ExecuteProvisionedProductPlan for more information about the operation and its arguments.

pub fn execute_provisioned_product_service_action(
&self
) -> ExecuteProvisionedProductServiceAction<C, M, R>

Constructs a fluent builder for the ExecuteProvisionedProductServiceAction operation.

See ExecuteProvisionedProductServiceAction for more information about the operation and its arguments.

pub fn get_aws_organizations_access_status(
&self
) -> GetAWSOrganizationsAccessStatus<C, M, R>

Constructs a fluent builder for the GetAWSOrganizationsAccessStatus operation.

See GetAWSOrganizationsAccessStatus for more information about the operation and its arguments.

pub fn get_provisioned_product_outputs(
&self
) -> GetProvisionedProductOutputs<C, M, R>

Constructs a fluent builder for the GetProvisionedProductOutputs operation.

See GetProvisionedProductOutputs for more information about the operation and its arguments.

pub fn import_as_provisioned_product(
&self
) -> ImportAsProvisionedProduct<C, M, R>

Constructs a fluent builder for the ImportAsProvisionedProduct operation.

See ImportAsProvisionedProduct for more information about the operation and its arguments.

pub fn list_accepted_portfolio_shares(
&self
) -> ListAcceptedPortfolioShares<C, M, R>

Constructs a fluent builder for the ListAcceptedPortfolioShares operation.

See ListAcceptedPortfolioShares for more information about the operation and its arguments.

pub fn list_budgets_for_resource(&self) -> ListBudgetsForResource<C, M, R>

Constructs a fluent builder for the ListBudgetsForResource operation.

See ListBudgetsForResource for more information about the operation and its arguments.

pub fn list_constraints_for_portfolio(
&self
) -> ListConstraintsForPortfolio<C, M, R>

Constructs a fluent builder for the ListConstraintsForPortfolio operation.

See ListConstraintsForPortfolio for more information about the operation and its arguments.

pub fn list_launch_paths(&self) -> ListLaunchPaths<C, M, R>

Constructs a fluent builder for the ListLaunchPaths operation.

See ListLaunchPaths for more information about the operation and its arguments.

pub fn list_organization_portfolio_access(
&self
) -> ListOrganizationPortfolioAccess<C, M, R>

Constructs a fluent builder for the ListOrganizationPortfolioAccess operation.

See ListOrganizationPortfolioAccess for more information about the operation and its arguments.

pub fn list_portfolio_access(&self) -> ListPortfolioAccess<C, M, R>

Constructs a fluent builder for the ListPortfolioAccess operation.

See ListPortfolioAccess for more information about the operation and its arguments.

pub fn list_portfolios(&self) -> ListPortfolios<C, M, R>

Constructs a fluent builder for the ListPortfolios operation.

See ListPortfolios for more information about the operation and its arguments.

pub fn list_portfolios_for_product(&self) -> ListPortfoliosForProduct<C, M, R>

Constructs a fluent builder for the ListPortfoliosForProduct operation.

See ListPortfoliosForProduct for more information about the operation and its arguments.

pub fn list_principals_for_portfolio(
&self
) -> ListPrincipalsForPortfolio<C, M, R>

Constructs a fluent builder for the ListPrincipalsForPortfolio operation.

See ListPrincipalsForPortfolio for more information about the operation and its arguments.

pub fn list_provisioned_product_plans(
&self
) -> ListProvisionedProductPlans<C, M, R>

Constructs a fluent builder for the ListProvisionedProductPlans operation.

See ListProvisionedProductPlans for more information about the operation and its arguments.

pub fn list_provisioning_artifacts(&self) -> ListProvisioningArtifacts<C, M, R>

Constructs a fluent builder for the ListProvisioningArtifacts operation.

See ListProvisioningArtifacts for more information about the operation and its arguments.

pub fn list_provisioning_artifacts_for_service_action(
&self
) -> ListProvisioningArtifactsForServiceAction<C, M, R>

Constructs a fluent builder for the ListProvisioningArtifactsForServiceAction operation.

See ListProvisioningArtifactsForServiceAction for more information about the operation and its arguments.

pub fn list_record_history(&self) -> ListRecordHistory<C, M, R>

Constructs a fluent builder for the ListRecordHistory operation.

See ListRecordHistory for more information about the operation and its arguments.

pub fn list_resources_for_tag_option(
&self
) -> ListResourcesForTagOption<C, M, R>

Constructs a fluent builder for the ListResourcesForTagOption operation.

See ListResourcesForTagOption for more information about the operation and its arguments.

pub fn list_service_actions(&self) -> ListServiceActions<C, M, R>

Constructs a fluent builder for the ListServiceActions operation.

See ListServiceActions for more information about the operation and its arguments.

pub fn list_service_actions_for_provisioning_artifact(
&self
) -> ListServiceActionsForProvisioningArtifact<C, M, R>

Constructs a fluent builder for the ListServiceActionsForProvisioningArtifact operation.

See ListServiceActionsForProvisioningArtifact for more information about the operation and its arguments.

pub fn list_stack_instances_for_provisioned_product(
&self
) -> ListStackInstancesForProvisionedProduct<C, M, R>

Constructs a fluent builder for the ListStackInstancesForProvisionedProduct operation.

See ListStackInstancesForProvisionedProduct for more information about the operation and its arguments.

pub fn list_tag_options(&self) -> ListTagOptions<C, M, R>

Constructs a fluent builder for the ListTagOptions operation.

See ListTagOptions for more information about the operation and its arguments.

pub fn provision_product(&self) -> ProvisionProduct<C, M, R>

Constructs a fluent builder for the ProvisionProduct operation.

See ProvisionProduct for more information about the operation and its arguments.

pub fn reject_portfolio_share(&self) -> RejectPortfolioShare<C, M, R>

Constructs a fluent builder for the RejectPortfolioShare operation.

See RejectPortfolioShare for more information about the operation and its arguments.

pub fn scan_provisioned_products(&self) -> ScanProvisionedProducts<C, M, R>

Constructs a fluent builder for the ScanProvisionedProducts operation.

See ScanProvisionedProducts for more information about the operation and its arguments.

pub fn search_products(&self) -> SearchProducts<C, M, R>

Constructs a fluent builder for the SearchProducts operation.

See SearchProducts for more information about the operation and its arguments.

pub fn search_products_as_admin(&self) -> SearchProductsAsAdmin<C, M, R>

Constructs a fluent builder for the SearchProductsAsAdmin operation.

See SearchProductsAsAdmin for more information about the operation and its arguments.

pub fn search_provisioned_products(&self) -> SearchProvisionedProducts<C, M, R>

Constructs a fluent builder for the SearchProvisionedProducts operation.

See SearchProvisionedProducts for more information about the operation and its arguments.

pub fn terminate_provisioned_product(
&self
) -> TerminateProvisionedProduct<C, M, R>

Constructs a fluent builder for the TerminateProvisionedProduct operation.

See TerminateProvisionedProduct for more information about the operation and its arguments.

pub fn update_constraint(&self) -> UpdateConstraint<C, M, R>

Constructs a fluent builder for the UpdateConstraint operation.

See UpdateConstraint for more information about the operation and its arguments.

pub fn update_portfolio(&self) -> UpdatePortfolio<C, M, R>

Constructs a fluent builder for the UpdatePortfolio operation.

See UpdatePortfolio for more information about the operation and its arguments.

pub fn update_portfolio_share(&self) -> UpdatePortfolioShare<C, M, R>

Constructs a fluent builder for the UpdatePortfolioShare operation.

See UpdatePortfolioShare for more information about the operation and its arguments.

pub fn update_product(&self) -> UpdateProduct<C, M, R>

Constructs a fluent builder for the UpdateProduct operation.

See UpdateProduct for more information about the operation and its arguments.

pub fn update_provisioned_product(&self) -> UpdateProvisionedProduct<C, M, R>

Constructs a fluent builder for the UpdateProvisionedProduct operation.

See UpdateProvisionedProduct for more information about the operation and its arguments.

pub fn update_provisioned_product_properties(
&self
) -> UpdateProvisionedProductProperties<C, M, R>

Constructs a fluent builder for the UpdateProvisionedProductProperties operation.

See UpdateProvisionedProductProperties for more information about the operation and its arguments.

pub fn update_provisioning_artifact(
&self
) -> UpdateProvisioningArtifact<C, M, R>

Constructs a fluent builder for the UpdateProvisioningArtifact operation.

See UpdateProvisioningArtifact for more information about the operation and its arguments.

pub fn update_service_action(&self) -> UpdateServiceAction<C, M, R>

Constructs a fluent builder for the UpdateServiceAction operation.

See UpdateServiceAction for more information about the operation and its arguments.

pub fn update_tag_option(&self) -> UpdateTagOption<C, M, R>

Constructs a fluent builder for the UpdateTagOption operation.

See UpdateTagOption for more information about the operation and its arguments.

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

Creates a client with the given service config and connector override.

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

Creates a new client from a shared config.

pub fn from_conf(conf: Config) -> Self

Creates a new client from the service Config.